Use Double for vital fields, fix specific device-link error messages
Backend team docs confirmed heartRate/oxygen/activity are decimal, not
integer -- switched the affected DTOs and Room entities from Int? to
Double?, added a formatVital() helper so whole-number values don't
render with a stray ".0", and added a destructive-migration fallback for
the schema change.

Also fixed the error-message extraction opencode added: it was parsing
a shape ({"errors": [{field, message}]}) that doesn't match the real API
error format ({"error": {"message", "errors": {"Code": ["Invalid code."]}}}),
so PairingRepository.linkDevice() was still showing the same generic
message for every 400 regardless of cause. Corrected the parser and
wired up the three documented cases (invalid/expired code, device
already registered).

+++ app/src/main/java/com/healthsync/app/data/local/PatientEntity.kt
@@ -12,9 +12,9 @@ data class PatientEntity(
     val notes: String?,
     val hasDevice: Boolean,
     val deviceSerialNumber: String?,
-    val lastHeartRate: Int?,
-    val lastOxygen: Int?,
-    val lastActivity: Int?,
+    val lastHeartRate: Double?,
+    val lastOxygen: Double?,
+    val lastActivity: Double?,
     val lastReadingAt: String?,
     val updatedAt: Long
 )
